# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIREST Sistema de Pacientes",
        "description": "Documentación de la API para gestionar el sistema pacientes",
        "contact": {
            "name": "Gustavo Galarza- Daniel Mamani",
            "email": "https://github.com/GustavoGalarza/sistema-pos.git"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/pacientes": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Obtener lista de pacientes",
                "description": "Devuelve todos los pacientes (Requiere autenticación)",
                "operationId": "735b2bb939a79d7a63ed2b9b004edef3",
                "responses": {
                    "200": {
                        "description": "Lista de pacientes obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "John"
                                            },
                                            "apellido": {
                                                "type": "string",
                                                "example": "Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "johndoe@example.com"
                                            },
                                            "telefono": {
                                                "type": "string",
                                                "example": "1234567890"
                                            },
                                            "fecha_nacimiento": {
                                                "type": "string",
                                                "example": "1990-01-01"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Crear un nuevo paciente",
                "description": "Crea un nuevo paciente (Requiere autenticación)",
                "operationId": "dc1a18dc26b35fd9e39b9e5a5ad74708",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "example": "1990-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paciente creado satisfactoriamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paciente creado satisfactoriamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/pacientesAll": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Obtener todos los pacientes",
                "description": "Devuelve una lista de todos los pacientes (Requiere autenticación)",
                "operationId": "edfe187ae042adb083a883de6d2329c5",
                "responses": {
                    "200": {
                        "description": "Pacientes encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "fecha_nacimiento": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "1990-01-01"
                                                    },
                                                    "telefono": {
                                                        "type": "string",
                                                        "example": "123456789"
                                                    },
                                                    "direccion": {
                                                        "type": "string",
                                                        "example": "Calle Falsa 123"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron pacientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontraron pacientes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}